+++ tests/core/full_node/stores/test_coin_store.py
@@ -117,7 +117,7 @@ async def test_basic_coin_store(db_version: int, softfork_height: uint32, bt: Bl
                     assert block.foliage_transaction_block is not None
                     await coin_store.new_block(
                         block.height,
-                        uint64(block.foliage_transaction_block.timestamp),
+                        block.foliage_transaction_block.timestamp,
                         block.get_included_reward_coins(),
                         tx_additions,
                         tx_removals,
@@ -127,7 +127,7 @@ async def test_basic_coin_store(db_version: int, softfork_height: uint32, bt: Bl
                         with pytest.raises(Exception):
                             await coin_store.new_block(
                                 block.height,
-                                uint64(block.foliage_transaction_block.timestamp),
+                                block.foliage_transaction_block.timestamp,
                                 block.get_included_reward_coins(),
                                 tx_additions,
                                 tx_removals,
@@ -185,7 +185,7 @@ async def test_set_spent(db_version: int, bt: BlockTools) -> None:
                         assert block.foliage_transaction_block is not None
                         await coin_store.new_block(
                             block.height,
-                            uint64(block.foliage_transaction_block.timestamp),
+                            block.foliage_transaction_block.timestamp,
                             block.get_included_reward_coins(),
                             additions,
                             removals,
@@ -233,7 +233,7 @@ async def test_num_unspent(bt: BlockTools, db_version: int) -> None:
                 additions: List[Coin] = []
                 await coin_store.new_block(
                     block.height,
-                    uint64(block.foliage_transaction_block.timestamp),
+                    block.foliage_transaction_block.timestamp,
                     block.get_included_reward_coins(),
                     additions,
                     removals,
@@ -265,7 +265,7 @@ async def test_rollback(db_version: int, bt: BlockTools) -> None:
                 assert block.foliage_transaction_block is not None
                 await coin_store.new_block(
                     block.height,
-                    uint64(block.foliage_transaction_block.timestamp),
+                    block.foliage_transaction_block.timestamp,
                     block.get_included_reward_coins(),
                     additions,
                     removals,
